[docker] Add an initial docker build
This patch adds an initial Docker file which defines a
multistage build that supports creating various Docker
images for Apache Kudu development.

This work is experimental and the choices and structure
may change with follow on commits.

Some things these images could be used for include:
- demos/examples
- pre-built binaries including thirdparty
- multi-os testing
- multi-os tools like build_mini_cluster_binaries.sh
- external integration testing

Some of the open tasks for the future are:
- Handle all build types.
- Optimize image sizes.
- Add kudu user to runtime images.
- Add an upload script with good tagging.
- Add healthchecks.

Change-Id: I95497b39e47f07301be75cbadd814656c7e2ea42
Reviewed-on: http://gerrit.cloudera.org:8080/12173
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>