KUDU-1868: Part 1: Add timer-based RPC timeouts
Currently, the Java client requires some kind of event to detect the
timeout of an RPC: either a response from the server in the chain of
sub-RPCs or a socket read timeout on the connection. This patch adds a
timer task to actively time out an RPC once it passes its deadline.

Part 2 will eliminate socket read timeouts from the Java client, except
possibly in the case of negotiation, which will fully resolve KUDU-1868.

There is one test included, which checks that timeouts occur without an
"outside stimulus" like a response from the server.

This patch should not degrade the performance of the client. Even though
every timer task holds a reference to its RPC, when the RPC completes it
cancels the timer task, which will make the timer release it at the next
tick. This means the RPC and its task should be available to be GC'd
after the next tick of the timer.

Change-Id: I8d823b63ac0a41cc5e42b63a7c19e0ef777e1dea
Reviewed-on: http://gerrit.cloudera.org:8080/12338
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>