[java] enable regression test for KUDU-2415
As KUDU-2463 fixed the bug in clean time advancement that now it is
guaranteed to have a non-zero clean time when servicing a scan, it is
safe to enable the regression test to ensure scanning a never-written-to
tablet from a fresh client with no propagated timestamp in
READ_YOUR_WRITE mode should work.

I looped TestScannerMultiTablet 1000 times and no failures are observed.

Change-Id: I9532a435c3b4cf2286630b6809ffa2c8ae1a4acd
Reviewed-on: http://gerrit.cloudera.org:8080/12278
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>