[java] deflake RYW tests in TestKuduClient
RYW tests in TestKuduClient is flaky which fails with AssertionError
when verifying the chosen snapshot timestamp returned from the server
is larger than the previous propagated timestamp before the scan.

It turns out to be a test only issue, which due to incorrect structure
of the test. It aims to test READ_YOUR_WRITE mode for multiple clients,
while the same client was being used. The distinction is important,
because we want to ensure a single client can still preserve
read-your-writes and read-your-reads session guarantees even there are
concurrent reads/writes performed by other clients in READ_YOUR_WRITE
scan mode.

Without the fix, 192/1000 runs of TestKuduClient failed with this error.
With the fix, 0/1000 runs of TestKuduClient failed.

Change-Id: I951abb9197f7e6b6a4c70cdf89948206840ddeda
Reviewed-on: http://gerrit.cloudera.org:8080/12276
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>