[thirdparty] detection of broken std::regex in libstdc++
Added detection of broken std::regex and friends in older
versions of libstdc++ (< 4.9) into the cpplint tool.  For details,
see [1].  I tried to find some warning flags for the GNU compiler
and/or linker to warn about non- or partially-implemented features
of GNU libstdc++ that doesn't yield compilation errors as is,
however I could not find anything relevant.

The motivation for this change is the fact that Kudu still compiles
and links against older libstdc++ at RHEL/CentOS 7, Ubuntu 14.04
and SLES 12, but using C++11 std::regex functionality could lead
to unexpected crash if the binary is linked aginst libstdc++ of
versions prior 4.9.  Yes, there are tests, but:
  * The issue with std::regex in older libstdc++ affects only the
    platforms mentioned above.
  * It's much better to detect and warn about the already known issue
    sooner to avoid wasting time on troubleshooting and debugging.

References:
  [1] https://gcc.gnu.org/onlinedocs/gcc-4.8.2/libstdc++/manual/manual/status.html#status.iso.2011
      section 28 'Regular Expressions'

Change-Id: Ibb7d01c536a3e2c059c9c9f6a8d14dcb6d6acad1
Reviewed-on: http://gerrit.cloudera.org:8080/12302
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins