KUDU-2665: LBM may delete containers with live blocks
Due to the nature of WritableBlock finalization/closing, it's
possible for a container with outstanding WritableBlock instances
to briefly appear as dead.
That's because:
1) The container's next block offset (responsible for determining
   fullness) is incrmented when the WritableBlock is finalized, but
2) The container's live block count is incremented when the
   WritableBlock is closed.
Thus, if the "last" block in a container is deleted after a
WritableBlock was finalized but before it has been closed, the
container will be erroneously marked as dead. When the container's
last referrent disappears, it will be deleted from disk despite
having live blocks in it.

Change-Id: I894f32b1c164ae7770c92171850edd167dfaf8ad
Reviewed-on: http://gerrit.cloudera.org:8080/12308
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>