[build] Fix mini cluster binary relocation on Mac OS
On Mac OS the ssl libs like `libssl.1.0.0.dylib` are read only.
When relocate_binaries_for_mini_cluster.py tries to copy
these read only libs, the build fails because the file
already exists and cannot be replaced.

This patch changes the copy_file method to stop matching
the permissions of the copied files. Additionally it clears the
artifact root on every run to ensure old files are not left behind.

Change-Id: Ib653b175ba2d0e5563509aecac372c9829eca2f5
Reviewed-on: http://gerrit.cloudera.org:8080/12310
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>