[java] fix ITClient test retrying
ITClient is flaky due to KUDU-2390, but it's also unable to retry upon
failure. I think this is caused by the static error latch which isn't reset
in between test runs. There's no reason for it to be static; I'm guessing
it's a holdover from when the minicluster state was static.

I also added an @After to close the client instance used by the test, which
will hopefully reduce the amount of log pollution in the retries.

Change-Id: I4a9f6a27541de02bbc8cf9dc3ba72824a6324a08
Reviewed-on: http://gerrit.cloudera.org:8080/12321
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>