KUDU-2670: Part 1: Build ScanToken by KeyRange
When reading data in a kudu table using spark,
if there is a large amount of data in the tablet,
reading the data takes a long time. The reason
is that KuduRDD uses a tablet to generate the
scanToken, so a spark task needs to process all
the data in a tablet.

We send SplitKeyRange RPC to TServer, split tablet's
primary key range into multiple primary key ranges
by size, and generate the scanToken by primary key ranges.

Change-Id: I0502f5d64569e8b1d45e88de3cb36aa2e01234d0
Reviewed-on: http://gerrit.cloudera.org:8080/12323
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>