[docker] Add a kudu user to the kudu image
As a best practice application images/containers should
not run as root if they don’t need to. This patch creates a
kudu user and group and creates/owns the kudu specific
directories.

I also added a tip to the readme about removing old tags.
I found myself needing to do this while working on this
patch.

Change-Id: Ia0d51cf8c6a585f98641de9b91d51e41dec9dbd0
Reviewed-on: http://gerrit.cloudera.org:8080/12344
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>