KUDU-2411: linux: Ship the sasl2 modules with the binary artifact
Because different Linux distributions put the dynamically-loaded
libsasl2 modules into different locations, we ship the libraries with
the binary artifact so that we can set SASL_PATH to a known good
location for them at runtime.

If we don't do this, we get a runtime error like the following:

Client connection negotiation failed: client connection to 127.25.108.252:41613: unable to find SASL plugin: PLAIN

We ran into ABI compatibility issues attempting to do this with macOS,
so for now we will rely on the version of libsasl2 shipped with the Mac
distribution.

This patch also contains a minor change to strip all recursively-found
files under lib/ so that the sasl2 modules will also be stripped at
build time on Linux.

Change-Id: I44afa05bcfa1a29a90f58c72ed6b0752014a16eb
Reviewed-on: http://gerrit.cloudera.org:8080/12350
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>