[sentry] Integrate AuthzProvider into CatalogManager
This commit enables master RPC authorization enforcement by connecting
the CatalogManager to the Sentry service via the SentryAuthzProvider.
When the Sentry integration is enabled (by setting the
--sentry_service_rpc_addresses flag), DDLs such as table creation,
alteration, deletion are validated to see if the connected user has
the permission to perform such operations. Note that the coarse-grained
access control is still applied to these endpoints. A --trusted_user_acl
flag is introduced to allow the trusted user, e.g. 'impala', to skip the
authorization enforcement.

Testing: This commit adds a new integration test (master_sentry-itest)
which tests that the integration works as expected with all exposed
table operations. More coverage on DDL stress tests with Sentry
integration enabled will be in a follow up patch.

Change-Id: Iab4aa027ae6eb4520db48ce348db552c9feec2a8
Reviewed-on: http://gerrit.cloudera.org:8080/11797
Tested-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>