[fs_manager-test] fix tests for macOS
As it turned out, it's been a while since the fs_manager-test started
failing.  In changelist b2f690d5 some changes were introduced
so one test started failing, and later on, with changelist 006937d3
the total number of failing tests on macOS grew up to three in total.

The reason behind the failures is simple: on macOS the only
block manager is FileBlockManager and it doesn't support many of
the operations exercised in the tests scenarios that started
failing.  This patch short-circuits those scenarios so they now
pass on macOS.

Change-Id: I19a8c6f48435bf5424500c86283490c9fed408f4
Reviewed-on: http://gerrit.cloudera.org:8080/12383
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins