KUDU-2411: Set SASL_PATH if needed when starting MiniCluster
When using the binary artifact on Linux, the MiniCluster should check to
see if the sasl2 libraries are included in the artifact and if should
set the SASL_PATH environment variable to the location of the
sasl2 modules when starting the MiniCluster. This is required because
the sasl2 modules are dynamically loaded at runtime.

Some refactoring was required to provide the relevant information to set
the SASL_PATH when starting the mini cluster process.

Additionally, this patch includes some miscellaneous cleanup.

Change-Id: Iaaeb30781f4483910c35a20c6d7c76f7f85aa4ce
Reviewed-on: http://gerrit.cloudera.org:8080/12351
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins