[rebalancer] always output summary on policy violations
This patch makes the location-aware rebalancer always output summary
on the detected violations of the replica placement policy.  The
detailed information on the violations is output if the runtime
flag --output_replica_distribution_details is specified.

Prior to this patch, the rebalancer would output either the summary
or the detailed information on the detected violations of the replica
placement policy, depending on the presence of the
--output_replica_distribution_details flag.

Change-Id: I5f68cebcef9a3c02203e80414082a0994a1721ee
Reviewed-on: http://gerrit.cloudera.org:8080/12358
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>