KUDU-1868 Part 2: Eliminate socket read timeout except for negotiation
This removes all use of socket read timeouts from the Java client and
deprecates all public APIs involving the socket read timeout. The Java
client does not use socket read timeouts anymore except to time out
negotiation. This is OK because negotiation is the only activity on the
socket when negotiation occurs. Once negotiation succeeds, the socket
read timeout handler is removed from the pipeline, and the timeout tasks
from Part 1 handle RPC timeouts.

The C++ client does not allow setting the negotiation timeout, and I
chose not to expose it in the Java client because it didn't seem useful,
so I set the timeout to 10000ms, which is much larger than the default
for the C++ client. The C++ client uses 3000ms, which is the server-side
default. The server-side default is also 3000ms, so I left some room for
the server-side default to rise. I wrote a test case and checked that
negotiation did time out when the constant was lowered to a small number
for the purposes of the test.

Change-Id: I391374dd72b6f4a91a9f69cf34758703afbdc59e
Reviewed-on: http://gerrit.cloudera.org:8080/12363
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins