Rename DeadlineTracker to TimeoutTracker
A deadline is the latest time by which something should be completed.
It's an instant, like "next Tuesday at noon". A timeout is an interval
of time allowed for some event to occur or be completed. It's a delta,
like "15 minutes". The DeadlineTracker tracked a "relative deadline",
relative to when the instance's "deadline" was set. That's actually a
timeout. This patch harmonizes the names to the concepts.

Change-Id: I3f465c925856390ecf4747e84bdd5a67c51c69eb
Reviewed-on: http://gerrit.cloudera.org:8080/12373
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>