Reduce "Failed to heartbeat..." log spam
Tablet servers heartbeat to the master at 1Hz by default, and previously
would log a message every time a heartbeat failed. This was a little
much, as it's not unusual for a server out of contact with a master to
fail heartbeats for quite some time. This patch lowers the frequency to
once per minute, and also enhances the error message to include the
number of failed heartbeats since the last successful heartbeat.

Change-Id: Ia48e844ebeb565343e55dbceae1674c34ea8ed5b
Reviewed-on: http://gerrit.cloudera.org:8080/12374
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins