[tools] KUDU-2688 rebalancer should not fix RF=2 violations
This patch addresses KUDU-2688.  Prior to this patch, the LA rebalancing
logic would suggest moves to reinstate the placement policy various
distributions of RF=2 tablet's replicas.  As a result, the rebalancer
looped indefinitely trying to implement the suggested replica movements.
However, RF=2 is a special case: any replica distribution of a tablet
with RF=2 is a violations of the placement policy constraints and it's
impossible to fix that, regardless of the number of locations
in a Kudu cluster.

With this patch, the rebalancer no longer tries to correct non-fixable
violations of the placement policy for RF=2.

Change-Id: I3afc7f3a228368e8646b46914ec7c31234a6aa85
Reviewed-on: http://gerrit.cloudera.org:8080/12375
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Kudu Jenkins