KUDU-2690: don't roll log schema on failed alter
It is possible to update the log segment header schema version based on
an AlterSchema operation that failed. This is because an alter operation
that didn't succeed (e.g. because of a schema version mismatch) is
treated as a successful transaction (similar to how we treat a
duplicated insert transaction as a successful transaction), and can thus
mistakenly lead to updating the log segment schema version, even on
failure. This can lead to a mismatch of schemas between the log segment
headers and the write ops in those log segments, which can lead to a
failure to bootstrap.

This patch addresses this by:
1. making the tablet no-op if it sees that an alter didn't go through,
2. storing the error in the commit message, so further bootstraps will
   skip over the op (this is KUDU-860).

Only the former is necessary to prevent the issue, and though the latter
uses extra space, it may be helpful for added visibility and debugging.

This patch adds a unit test that reproduced the scenario at the
TabletReplica level, as well as a less targeted test that uses the same
principal to test that we update the log segment schema upon replaying
an alter. A more end-to-end test that arrives at the reported state by
working around master-level table locking can be found here[1].

[1] https://gist.github.com/andrwng/3a049bb038680cc0254c5ba52b9a7507

Change-Id: Id761851741297e29a4666bec0c34fc4f7285f715
Reviewed-on: http://gerrit.cloudera.org:8080/12462
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>