Schema copy and move constructors/operators must copy max_col_id_
The CopyFrom() helper method used by the Schema copy constructor and
assignment operator, as well as the move constructor and move operator,
previously did not copy the value of max_col_id_, leaving it
uninitialized in the copy. This patch addresses that oversight.

This patch also updates the Schema::Equals() method to provide the
option (not used by default) to check that the column ids and max column
id match between the compared schemas, allowing us to provide coverage
by beefing up an existing unit test.

Change-Id: I61b8df1142dc733c446a5717ca8cae1309f3e867
Reviewed-on: http://gerrit.cloudera.org:8080/12415
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>