[spark-tools] DistributedDataGenerator repartition support
This patch adds support to the DistributedDataGenerator
to repartition the data to match the Kudu partitioning.

Because data generation is now decoupled from data
loading, this patch changes the collision handling
behavior. Instead of generating new data on collision,
now the collision is only tracked in the metrics.

Additionally this patch changes the default generation
type from random to sequential given that has been
shown to be the more common option and the type
of workload Kudu is better suited for.

Change-Id: I57bcc68d645c52b429ac6cf8bcdf0551a8244995
Reviewed-on: http://gerrit.cloudera.org:8080/12411
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>