KUDU-1900: add loopback check and test
This change modifies Socket::IsLoopbackConnection to check
if remote IP is in 127.0.0.0/8 subnet. With this change
all connections from 127.0.0.0/8 are treated as local.

This change adds force_external_client_ip parameter to
AuthTokenIssuingTest. Before this change, AuthTokenIssuingTest
relied on Kudu treating connections between non-matching
addresses in 127.0.0.0/8 subnet as external. With this change,
the test forces Kudu client to use an non-loopback IP
address for test cases that verify external connections.
If an external IP address is not available, test cases that
require an external IP will be skipped.

For convenience, this change adds two static utility methods
(HostPort::IsLoopback and HostPort::AddrToString), and refactors
Sockaddr::host and Sockaddr::IsAnyLocalAddress to use these
static methods.

Change-Id: I3483a9729ddeeb7901e3738532a45b49e713208f
Reviewed-on: http://gerrit.cloudera.org:8080/12474
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins