[tools] Add tool to list data dirs for a tablet
It's hard to find out what the data dirs are for a tablet, given its
tablet id, but it might be useful to do so. For example, if a user is
trying to correlate disk utilization to busy tablets, it's useful to
know which tablets will cause disk activity on which devices.

Change-Id: I6c6de565e4061eb8248b453d2230847e064040ff
Reviewed-on: http://gerrit.cloudera.org:8080/12417
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>