KUDU-2411: Update license information for binary test jars
This patch updates the license files to account for transitive
dependencies we ship with the binary jar files and updates the scripting
infrastructure to validate the licensing information at binary jar build
time.

This patch also introduces binary-jar-specific license and notice files
to account for licensing information only relevant to the binary
artifacts.

Change-Id: If864384e60013d592c5c7315e2d9a99c83f31449
Reviewed-on: http://gerrit.cloudera.org:8080/12431
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins