[tools] Support running the master and tablet server via the kudu binary
Adds commands to the kudu tools to run the master
and tablet server:
   - `kudu master run ...`
   - `kudu tserver run ...`

This means we can ship a single binary in the kudu
docker image or potentially the kudu-binary jar
reducing the size by approximately 66%. Though
follow up changes may be needed in the kudu-binary
case.

The behavior is the same as running the kudu-master
and kudu-tserver binaries. The logging initialization was
updated to ensure that the logging filenames are
the same as when running via the dedicated binaries.
A few logging fixes are included to ensure this capability
is possible.

This patch also contains a few related fixes for default
flags ensuring that they can be overwritten when
necessary. These changes also ensure flags are not
incorrectly reported as non-default.

Change-Id: I3717cbac930b3506a76f7a51388c64afbcbb480e
Reviewed-on: http://gerrit.cloudera.org:8080/12517
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>