Add GetTableLocations tracing
This adds minimal tracing to GetTableLocations. It's useful because
adding a trace event for the call makes it easier to count the calls in
bulk in a trace, and having the requestor string traced makes it
possible to attribute GetTableLocations calls to specific clients in
many cases, given that the trace will also include the hostport of the
client.

Change-Id: I92e0dd18de52596f2b9a419c1b6cf7a5666d24bc
Reviewed-on: http://gerrit.cloudera.org:8080/12585
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>