KUDU-2710: Fix KeepAliveRequest retries
Fixes KeepAliveRequest retries by adding a partitionKey
implementation. Without this a null partitionKey is passed
and the client treats this as a master table.

A follow up patch should include fixes to prevent issues
like this in the future and fix any remaining retry issues.
This patch is kept small to ensure easy backports.

Change-Id: I951212ab76079e5788c2870223b45782b16509e7
Reviewed-on: http://gerrit.cloudera.org:8080/12586
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>