KUDU-2710: Fix KeepAliveRequest retries
Fixes KeepAliveRequest retries by adding a partitionKey
implementation. Without this a null partitionKey is passed
and the client treats this as a master table.

A follow up patch should include fixes to prevent issues
like this in the future and fix any remaining retry issues.
This patch is kept small to ensure easy backports.

Change-Id: I573c1b6a14a4277c0f837c189175ed91e4e9e933
Reviewed-on: http://gerrit.cloudera.org:8080/12606
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins