[docs] Remove outdated package and quickstart references
The Cloudera Quickstart VM has not been updated since
CDH 5.13/Kudu 1.5.0. Addtionally the demo repository which
contained the bootstrap.sh script has been retired. All
references to the quickstart VM are removed and an Apache
specific alternative for getting started should be added back
in a follow on patch.

The packages install was also outdated as it referenced
a package repository which was last updated for Kudu 1.4.0.
This is because Cloudera started packaging Kudu with CDH
after that version. Being able to install via packages is useful
and a reference to the Cloudera documentation could be
added in an appropriate location in a follow on patch.
Alternatively, Apache Kudu packages could be created.

While searching for references to the packages and
quickstart I also saw a few outdated Cloudera references
and updated them.

Change-Id: I3a8025c2308990cdf16205aca03bdb0e99c6cbe7
Reviewed-on: http://gerrit.cloudera.org:8080/12590
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Grant Henke <granthenke@apache.org>