schema: Fix ColumnSchema::Equals() usage
In commit 4be39feef4907550c16f9d2b34430d098b9f1b21 the signature of
ColumnSchema::Equals() was changed from taking a boolean flag for
checking whether the defaults matched to taking an enum for describing a
broader set of comparison options. However, a few call sites were not
updated when that change was made, and due to type coercion several call
sites ended up with weaker comparison semantics.

This patch reinstates the behavior prior to the merge of the
above-mentioned patch for the problematic call sites and updates the
function signature of ColumnSchema::Equals() to treat the comparison
enum as strongly typed to avoid similar problems in the future.

Updated tests to add coverage for the relevant call sites.

Change-Id: I416e4a44d0ed259d251d52a663b72acd79b46f33
Reviewed-on: http://gerrit.cloudera.org:8080/12605
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins